requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pypdf2>=3.0.0